# trace: remove contention on trace ID generation