#### Summary
Use a `sync.Pool[*rand.Rand]` for generating `int64` trace IDs instead of the global `math/rand.Int63()`, which grabs a global mutex.  This mutex showed up as having high contention and limiting throughput in an internal benchmark.

On a 4 CPU machine, this reduces the runtime of StartChildSpan by 74%:
```
name              old time/op    new time/op    delta
StartChildSpan      72.1ns ± 1%    80.9ns ± 1%  +12.16%  (p=0.002 n=6+6)
StartChildSpan-2    92.9ns ± 1%    48.2ns ± 2%  -48.07%  (p=0.002 n=6+6)
StartChildSpan-4     125ns ± 0%      35ns ± 2%  -72.39%  (p=0.004 n=5+6)
```

#### Motivation
Improved scalability for HTTP servers that generate trace IDs for each request.

#### Test plan
Added unit tests for the new functionality, and ensured existing tests pass for the `trace` changes.

#### Rollout/monitoring/revert plan
`go get github.com/stripe/veneur@latest` in downstream repos and ensure tests pass and benchmarks improve